refactor: reorganize components and update routing structure
