+++ src/Root.tsx
@@ -1,10 +1,8 @@
 import { HashRouter, Routes, Route, Navigate } from 'react-router-dom';
 import { App } from './App';
-import HomePage from './components/HomePage';
-import TabsPage from './components/TabsPage';
-import TabItemPage from './components/TabItemPage';
-import NotFoundPage from './components/NotFoundPage';
-import PleaseSelectTab from './components/PleaseSelectTab';
+import HomePage from './pages/HomePage';
+import TabsPage from './pages/TabsPage';
+import NotFoundPage from './pages/NotFoundPage';
 
 export const Root = () => (
   <HashRouter>
@@ -12,9 +10,9 @@ export const Root = () => (
       <Route path="/" element={<App />}>
         <Route index element={<HomePage />} />
         <Route path="home" element={<Navigate to="/" replace />} />
-        <Route path="tabs" element={<TabsPage />}>
-          <Route index element={<PleaseSelectTab />} />
-          <Route path={`:tabId`} element={<TabItemPage />} />
+        <Route path="tabs">
+          <Route index element={<TabsPage />} />
+          <Route path={`:tabId?`} element={<TabsPage />} />
         </Route>
         <Route path="*" element={<NotFoundPage />} />
       </Route>
