--- src/components/PleaseSelectTab.tsx
@@ -1,5 +1,0 @@
-const PleaseSelectTab = () => {
-  return <p>Please select a tab</p>;
-};
-
-export default PleaseSelectTab;
